HOLO-54 <chore: 파일명 수정>
