+++ src/app/trash-guide/_components/WasteDaySelector.tsx
@@ -1,20 +1,14 @@
 "use client";
-
-import type { RegionData, WasteDayAnswerData } from "@/app/trash/_types/day";
-import { useQuery } from "@tanstack/react-query";
-
-import { fetchOpenAiDay } from "@/app/trash/_actions/actions";
+import { fetchOpenAiDay } from "@/app/trash-guide/_actions/actions";
+import type { WasteDayAnswerData } from "@/app/trash-guide/_types/day";
+import RegionSelect from "@/components/common/RegionSelect";
 import { clsx } from "clsx";
 import { useState } from "react";
-
-const fetchRegionData = async (): Promise<RegionData[]> => {
-  const response = await fetch("data/district.json");
-  return response.json();
-};
 
 const WasteDaySelector = () => {
   const [regionSelected, setRegionSelected] = useState<string>("");
   const [districtSelected, setDistrictSelected] = useState<string>("");
+  const [districts, setDistricts] = useState<string[]>([]);
   const [wasteDayAnswer, setWasteDayAnswer] =
     useState<WasteDayAnswerData>(null);
   const [loading, setLoading] = useState<boolean>(false);
@@ -28,36 +22,24 @@ const WasteDaySelector = () => {
       }
     } catch (error) {
       console.error(error);
+      throw Error("쓰레기 배출 요일을 알려주는 OpenAI 오류가 발생했습니다.");
     } finally {
       setLoading(false);
     }
   };
 
-  const {
-    data: regionData,
-    isPending,
-    error,
-  } = useQuery<RegionData[]>({
-    queryKey: ["regionData"],
-    queryFn: fetchRegionData,
-  });
-
-  const handleChangeRegion = (
-    e: React.ChangeEvent<HTMLSelectElement>,
-  ): void => {
-    setRegionSelected(e.target.value);
+  const handleRegionChange = (region: string, districts: string[]) => {
+    setRegionSelected(region);
     setDistrictSelected("");
+    setDistricts(districts);
   };
 
-  //TODO: 버튼 스타일 바꾸기
   const btnDisabled = clsx("border border-yellow-500", {
-    // 활성화시
     "bg-yellow-500":
       regionSelected &&
       regionSelected !== "지역선택" &&
       districtSelected &&
       districtSelected !== "시군구",
-    // 비활성화시
     "bg-gray-400":
       !regionSelected ||
       regionSelected === "지역선택" ||
@@ -65,28 +47,15 @@ const WasteDaySelector = () => {
       districtSelected === "시군구",
   });
 
-  if (isPending) return <div>로딩중</div>; // TODO: 로딩 바꾸기
-  if (error instanceof Error) return <div>에러: {error.message}</div>; // TODO: 에러처리 바꾸기
-
   return (
     <div>
       <p>지역별 쓰레기 배출 요일 확인</p>
       <div>
         <p>지역선택</p>
-        <select
-          name="region"
-          id="region"
-          className="border border-yellow-500"
-          onChange={handleChangeRegion}
-          value={regionSelected}
-        >
-          <option value="지역선택">지역선택</option>
-          {regionData?.map((region, i) => (
-            <option value={Object.keys(region)[0]} key={i}>
-              {Object.keys(region)[0]}
-            </option>
-          ))}
-        </select>
+        <RegionSelect
+          selectedRegion={regionSelected}
+          onChange={handleRegionChange}
+        />
         <select
           name="district"
           id="district"
@@ -96,18 +65,11 @@ const WasteDaySelector = () => {
           disabled={!regionSelected}
         >
           <option value="시군구">시군구(전체) 선택</option>
-          {regionData
-            ?.find((region) => Object.keys(region)[0] === regionSelected) // 선택한 지역을 data에서 찾기
-            ?.[regionSelected]?.map(
-              (
-                district: string,
-                i: number, // 찾은 지역의 시군구 목록을 옵션으로 만들기
-              ) => (
-                <option value={district} key={i}>
-                  {district}
-                </option>
-              ),
-            )}
+          {districts.map((district, i) => (
+            <option value={district} key={i}>
+              {district}
+            </option>
+          ))}
         </select>
         <button
           type="button"
@@ -123,9 +85,7 @@ const WasteDaySelector = () => {
           조회
         </button>
       </div>
-      {/* TODO: ai 답변전 텍스트 수정 */}
       {!wasteDayAnswer && !loading && <div>지역을 선택해주세요.</div>}
-      {/* TODO: 로딩 수정 */}
       {loading && <div>쓰레기 배출 요일을 확인중입니다...</div>}
       <ul>
         {wasteDayAnswer?.map((answer, i) => (
