+++ src/app/trash-guide/_types/day.ts
