+++ src/app/trash-guide/page.tsx
@@ -1,0 +1,20 @@
+import FoodWasteCheck from "@/app/trash-guide/_components/FoodWasteCheck";
+import WasteDaySelector from "@/app/trash-guide/_components/WasteDaySelector";
+
+const page = () => {
+  return (
+    <div>
+      <div>
+        <h2>쓰레기 가이드</h2>
+        <p>우리 동네 쓰레기 배출 일정과 분리배출 정보</p>
+      </div>
+      <div className="mx-auto max-w-xl border border-yellow-500">
+        <WasteDaySelector />
+        <hr className="my-5" /> {/* TODO: 이후 구분선삭제 */}
+        <FoodWasteCheck />
+      </div>
+    </div>
+  );
+};
+
+export default page;
